+++ src/components/CopyUiProvider/copy-ui-provider.tsx
@@ -1,8 +1,11 @@
+import { MDXProvider } from "@mdx-js/react";
 import { createContext, type FC, useEffect, useRef, useState } from "react";
 import { createStore, type StoreApi } from "zustand";
 
 import { Toast } from "@/components/Toast";
+import { Typography } from "@/components/Typography";
 import { createJssState, type JssState } from "@/utils/jss";
+
 import { DEFAULT_THEME } from "./default-theme";
 import type { Theme } from "./theme";
 import { useJss } from "./use-jss";
@@ -31,10 +34,17 @@ const CopyUiProvider: FC<CopyUiProviderProps> = (props) => {
     }));
   }
 
+  const mdxComponents = {
+    h1: Typography.H1,
+    h2: Typography.H2,
+    h3: Typography.H3,
+    h4: Typography.H4,
+  };
+
   return (
     <CopyUiContext.Provider value={storeRef.current}>
       <Toast.Context>
-        {children}
+        <MDXProvider components={mdxComponents}>{children}</MDXProvider>
         <Toast.Container />
       </Toast.Context>
       <JssSsrCss />