+++ src/components/CopyUiProvider/index.ts
@@ -1,4 +1,4 @@
-// CopyUiProvider from copy-ui @ 2025-11-06.
+// CopyUiProvider from copy-ui @ 2025-11-08.
 
 export { CopyUiProvider } from "./copy-ui-provider";
 export { useJss } from "./use-jss";