+++ vite.config.ts
@@ -1,10 +1,24 @@
 import { reactRouter } from "@react-router/dev/vite";
 import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
+import mdx from '@mdx-js/rollup';
+import path from "node:path";
 
 export default defineConfig({
   build: {
     target: ["chrome107", "edge107", "firefox104", "safari16"],
   },
-  plugins: [reactRouter(), tsconfigPaths()],
+  plugins: [
+    {
+      enforce: 'pre',
+      ...mdx({ providerImportSource: '@mdx-js/react', include: /.*\.mdx$/ }),
+    },
+    reactRouter(),
+    tsconfigPaths(),
+  ],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
+  },
 });